# 05/results/api.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "RESTful API Design"
  description: "Simple design of a company that runs an application for a travel agency."
schemes:
- "https"
host: "t-a.org"
...
paths:
...
  /tour:
    get:
      parameters:
        - name: "customer"
          in: "query"
          type: "string"
          description: "list all tours in queried customer"
      tags:
        - "tour"
      summary: "list all tours"
      description: ""
      operationId: "geTours"
      responses:
        "200":
          description: "listing tours"
          schema:
            $ref: '#/definitions/tour'
    post:
      tags:
        - "tour"
      summary: "add new tour"
      description: ""
      operationId: "postTour"
      parameters:
        - in: "body"
          name: "body"
          description: "specification of a new tour object"
          required: true
          schema:
            $ref: "#/definitions/tour"
      responses:
        "200":
          description: "location succesfully added"
          schema:
            $ref: '#/definitions/tour'
        "405":
          description: "invalid input"
  /tour/{tourid}:
    get:   
      tags:
        - "tour"
      summary: "list tour"
      description: ""
      operationId: "getTour"
      parameters:
        - name: "tourid"
          in: "path"
          description: "id of tour that needs to be updated"
          required: true
          type: "string"
      responses:
        "200":
          description: "listing tour"
          schema:
            $ref: '#/definitions/tour'
    put:
      tags:
        - "tour"
      summary: "update tour"
      description: ""
      operationId: "putTour"
      parameters:
        - name: "tourid"
          in: "path"
          description: "id of tour that needs to be updated"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "specification of an updated tour object"
          required: true
          schema:
            $ref: "#/definitions/tour"
      responses:
        "200":
          description: "tour succesfully updated"
        "400":
          description: "invalid id"
        "404":
          description: "tour not found"
    delete:
      tags:
        - "tour"
      summary: "delete tour"
      description: ""
      operationId: "deleteTour"
      parameters:
        - name: "tourid"
          in: "path"
          description: "id of tour that needs to be deleted"
          required: true
          type: "string"
      responses:
        "200":
          description: "tour succesfully deleted"
        "400":
          description: "invalid id"
        "404":
          description: "tour not found"
...
definitions:
  country:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
  location:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      country:
        type: "object"
        format: "#/definitions/country"
  tour:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      location:
        type: "object"
        format: "#/definitions/location"
      customer:
        type: "object"
        format: "#/definitions/customer"
  customer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"